more work, should probably delete a lot

+++ src/main/java/xyz/bluspring/kilt/injections/world/inventory/AnvilMenuInjection.java
@@ -1,7 +1,7 @@
 package xyz.bluspring.kilt.injections.world.inventory;
 
 public interface AnvilMenuInjection {
-    default void setMaximumCost(int value) {
+    default void setMaximumCost(long value) {
         throw new IllegalStateException();
     }
 }
